Initial uploads, server and world files
The first upload of the files needed to run a server, as well as
incomplete game classes. Most skeleton code is written in the GameWorld
file; however, no actual logic is completed as of yet. The Server and
ServerThread files run, but are not final.  TextUtils provides support
for loading and saving GameWorld files, but the specific implementation
will likely change as time goes on.