TextUtils added back, new exception class
The TextUtils class was re-created, GameWorld was getting too large with
the save and load helper methods. The implementation of save and load
methods was changed to provide a less error-prone method. All save files
are now written to a single, constant path, which should be determined
by the Server on startup.